@@ -1,6 +1,6 @@
 package org.mercuriete.musiciantools.bpmcalculator
 
-import android.support.annotation.Nullable
+import androidx.annotation.Nullable
 import javax.inject.Inject
 
 class BPMCalculatorPresenter @Inject constructor(private var bpmCalculatorService: BPMCalculatorService) : BPMCalculatorContract.Presenter {
